REST commands to cluster with no region is successful
