+++ geode-assembly/src/distributedTest/java/org/apache/geode/rest/internal/web/controllers/RestAPICompatibilityTest.java
@@ -15,27 +15,14 @@
  */
 package org.apache.geode.rest.internal.web.controllers;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.apache.geode.rest.internal.web.controllers.RestUtilities.executeAndValidateGETRESTCalls;
+import static org.apache.geode.rest.internal.web.controllers.RestUtilities.executeAndValidatePOSTRESTCalls;
 
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -43,7 +30,6 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import org.apache.geode.cache.RegionShortcut;
-import org.apache.geode.management.operation.RebalanceOperation;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
 import org.apache.geode.test.dunit.rules.MemberVM;
 import org.apache.geode.test.junit.categories.BackwardCompatibilityTest;
@@ -74,20 +60,9 @@ public class RestAPICompatibilityTest {
 
   public RestAPICompatibilityTest(String oldVersion) throws JsonProcessingException {
     this.oldVersion = oldVersion;
-    postRESTAPICalls = new HashMap<>();
-    // {REST endpoint,{Body, Successful Status Message, Introduced in version}}
-    postRESTAPICalls.put("/management/v1/operations/rebalances",
-        new String[] {mapper.writeValueAsString(new RebalanceOperation()), "Operation started",
-            "1.11.0"});
+
   }
 
-  private static Map<String, String[]> postRESTAPICalls;
-
-
-  private static final String[][] getRESTAPICalls = {
-      // REST endpoint , status
-      {"/geode-mgmt/v1/management/commands?cmd=rebalance", "OK"}
-  };
 
   @Test
   public void restCommandExecutedOnLatestLocatorShouldBeBackwardsCompatible() throws Exception {
@@ -125,70 +100,8 @@ public class RestAPICompatibilityTest {
     // Execute REST api calls to from the new locators to the old servers to ensure that backwards
     // compatibility is maintained
 
-    executeAndValidatePOSTRESTCalls(locator1.getHttpPort());
+    executeAndValidatePOSTRESTCalls(locator1.getHttpPort(), oldVersion);
     executeAndValidateGETRESTCalls(locator1.getHttpPort());
 
-  }
-
-  void executeAndValidatePOSTRESTCalls(int locator) throws Exception {
-
-    try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
-      for (Map.Entry<String, String[]> entry : postRESTAPICalls.entrySet()) {
-        // Skip the test is the version is before the REST api was introduced.
-        if (TestVersion.compare(oldVersion, entry.getValue()[2]) < 0) {
-          continue;
-        }
-        HttpPost post =
-            new HttpPost("http://localhost:" + locator + entry.getKey());
-        post.addHeader("Content-Type", "application/json");
-        post.addHeader("Accept", "application/json");
-        StringEntity jsonStringEntity =
-            new StringEntity(entry.getValue()[0], ContentType.DEFAULT_TEXT);
-        post.setEntity(jsonStringEntity);
-        CloseableHttpResponse response = httpClient.execute(post);
-
-        HttpEntity entity = response.getEntity();
-        InputStream content = entity.getContent();
-
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(content))) {
-          String line;
-          StringBuilder sb = new StringBuilder();
-          while ((line = reader.readLine()) != null) {
-            sb.append(line);
-          }
-          JsonNode jsonObject = mapper.readTree(sb.toString());
-          String statusCode = jsonObject.findValue("statusCode").textValue();
-          assertThat(statusCode).satisfiesAnyOf(
-              value -> assertThat(value).isEqualTo("ACCEPTED"),
-              value -> assertThat(value).contains("OK"));
-          String statusMessage = jsonObject.findValue("statusMessage").textValue();
-          assertThat(statusMessage).contains(entry.getValue()[1]);
-        }
-      }
-    }
-  }
-
-  public static void executeAndValidateGETRESTCalls(int locator) throws Exception {
-
-    try (CloseableHttpClient httpclient = HttpClients.createDefault()) {
-      for (String[] commandExpectedResponsePair : getRESTAPICalls) {
-        HttpGet get =
-            new HttpGet("http://localhost:" + locator +
-                commandExpectedResponsePair[0]);
-        CloseableHttpResponse response = httpclient.execute(get);
-        HttpEntity entity = response.getEntity();
-        InputStream content = entity.getContent();
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(content))) {
-          String line;
-          StringBuilder sb = new StringBuilder();
-          while ((line = reader.readLine()) != null) {
-            sb.append(line);
-          }
-          JsonNode jsonObject = mapper.readTree(sb.toString());
-          String statusCode = jsonObject.findValue("status").textValue();
-          assertThat(statusCode).contains(commandExpectedResponsePair[1]);
-        }
-      }
-    }
   }
 }
